fix(bfs): use Float64Array for cost to preserve staleness-check equality
The cultures + states BFS rewrites used Float32Array for the cost array
while the queue stores Float64 priorities. Computed totalCost values
(priority + 10 + cellCost / expansionism) almost always have fractional
parts that aren't exactly representable in Float32.

Sequence of the bug:
1. compute totalCost as Float64
2. cost[en] = totalCost  → silently rounded to nearest Float32
3. queue.push(en, totalCost)  → original Float64 stored in queue
4. on pop: priority (Float64) !== cost[e] (Float32 rounded back) → MISMATCH
5. staleness check skips the entry, BFS never processes the cell's neighbours

Effect: state/culture is assigned at push (line `cellsState[en] = s`), so
capitals + first-hop neighbours appear, but the BFS terminates at depth 1
and never expands further. Visible at 10K cells as completely uncoloured
state regions despite 16 states existing.

Float64 the cost array. Memory cost is negligible (8 bytes/cell vs 4 —
1.6MB at 200K cells).

+++ src/modules/cultures-generator.ts
@@ -1237,7 +1237,9 @@ class CulturesModule {
 
     const n = cells.i.length;
     const queue = new FlatQueue();
-    const cost = new Float32Array(n);
+    // Float64Array — Float32 rounding breaks the staleness check (see equivalent
+    // comment in states-generator.ts). BFS would stop at first hop.
+    const cost = new Float64Array(n);
     const cellCulture = new Uint16Array(n);
 
     const neutralRate = (document.getElementById("neutralRate") as HTMLInputElement | null)?.valueAsNumber || 1;
